threadpool: use atomic compare and swap when consuming rows
The GCC __sync_and_and_fetch() instrinsic was returning the new value, which
is entirely useless.  The thread could not tell if it was the one which cleared
the bit.  So I've switched it to a compare and swap operation.  Now the thread
pool test runs correctly when compiled with GCC