TEncSearch: clip mvmax by a fixed ammount for frame parallelism
When frame parallelism is enabled, we enforce at least merange + interpolation
half-length of reconstructed reference pixels to be available in all reference
frames before allowing the encoder to begin compressing a given row.  So we
cannot allow the motion search to extend beyond merange past the colocated ref
block (regardless of the MVP search starting location).

We can't use the actual number of rows available because this results in non-
determinism.