aq: fixes for loop over 16x16 blocks
This loop was busted when maxCUSize was not 64.  It still has a problem with
pictures that are not even multiples of 16.  The lookahead will extend out the
frame during lowres init to an even multiple of 16 pixels, so it's lowres CU
width will be wider than the AQ code will use, so the block_xy offsets will be
wrong for lookahead analysis.

The pixel extension needs to be moved earlier so AQ and the lookahead have a
consistent 16x16 CU width