refactor: move motion reference pointers from TEncSearch to Slice
The actual storage for the MotionReference instances lives in the FrameEncoder,
to save memory, but we don't want the analysis code to be aware of the FE so
we give it a pointer to this array.  Having that pointer live in TEncSearch
forces us to update this pointer each time a worker thread starts a new job.

Having the pointer in Slice means we only have to set it once, when the slice
first starts to be encoded. And having this pointer next to the RPS data in the
slice header makes this more coherent.