Make FrameEncoder partially virtual so it can be overloaded
FrameEncoder is a logical place to use a hardware accelerator or an alternative core encoder.
By making a few function virtual, it can be easily replaced by an inheriting class, transparently for the Encoder.
This way all the RC/interaction code can stay within the FrameEncoder class and reduce conflicts in later updates.
---
 source/encoder/encoder.cpp    | 32 ++++++++++++++++++--------------
 source/encoder/encoder.h      |  2 +-
 source/encoder/frameencoder.h |  8 ++++----
 3 files changed, 23 insertions(+), 19 deletions(-)