cmake: remove buggy workarounds for partial SIMD support (fixes #92)
In the past, there were a number of primitives written in SIMD intrinsics that
could work without compiling with YASM. Most of those are now gone, and we
generally require YASM for SIMD support. This commit remoes support for using
the few remaining SIMD intrinsics without having YASM to provide
implementations of x265_emms(), x265_cpu_cpuid(), etc. Fixing a bug in the
process.