motion: add a version of setSourcePU which can accept fenc from another Yuv
The analysis code has already gone through the trouble of loading the CU's fenc
pixels from the source picture into a much smaller Yuv buffer with small
strides. This allows us to avoid accessing the fenc PicYuv in a performance
critical portion of the encoder.

We utilize the Yuv class to copy the PU, since it already has logic for
calculating part offsets for luma and chroma