slicetype: allow queue to fill past full to prevent bottlenecks
Allow the lookahead to grow just past full before we begin pulling off output
frames and handing them to frame encoders.  This lag of about one mini-gop
allows slicetypeDecide to stay ahead of the frame encoders and always have
frames in the output queue when they are needed.  It's a non-trivial performance
boost for most presets that used b-adapt 2.