stats: introduce X265_LOG_FRAME for file level CSV logging without console logs
Using --log-level debug to trigger frame level CSV logging is problematic since
the console logging is often a big enough overhead that it influences the
performance characteristics. --log-level frame will log frame level stats to
the CSV without enabling frame-level console logging. Note that this does not
change the behavior of --log-level debug, but it does change the behavior of
--log-level 3.