param: no longer disable deblock at ultrafast
Recent profiling shows that it has an out-sized effect on quality for the very
small performance impact.

Deblocking accounts for about 3% of the CPU time at ultrafast preset, but at
smaller resolutions it has an outsised effect on frame parallelism because
deblocking introduces a 1-row reference lag.

At 4k this lag is not important, at 1080p the lag becomes measurable, while at
720p the lag can account for about a %10 drop in performance.