search: make bMergeOnly argument to predInterSearch more of a suggestion
predInterSearch() will only check merge candidates for PUs when bMergeOnly is
true, and currently returns false when no merge candidates are available. This
changes the logic to simply fall into uni-directional and bi-directional motion
estimation when no merge candidates are available and ignore the 'merge-only'
flag in these cases.

bMergeOnly is only used for AMP modes with 64x64 partitions, a performance
optimization from the HM which we have kept. It is perfectly valid to use ME for
these partitions, and better to have predInterSearch() always return an inter
mode rather than having to handle rare error conditions.

This might change outputs in minor ways when AMP is enabled