test: add lists of regression tests to be used by external test harness
The test harness is being tracked externally in
https://bitbucket.org/sborho/test-harness for two reasons:

1 - we want to be able to use the latest harness features on older versions of
    the encoder

2 - output-changing-commits.txt must be maintained outside the source repository
    since it contains changeset hashes. If it was tracked inside the source repo
    it would always get updated after output changing commits which would cause
    problems when you checked out an output changing commit for test, you would
    not know that the next commit would mark it as such.

That said, we do want to keep the list of regression tests within the repository
itself so new tests can be added as new comamnd line params are introduced. Or
params changed, etc. By tracking these lists within the repository, the lists
are always appropriate for the revision being tested (after these commits,
anyway)