asm: interp_8tap_vert_pX sse2
This code replaces c code for sse2.  It is the combination of the sse4 macros into
one for smaller code size with no sacrifice in function and a few tweeks for performance.
The original sse4 macros only use up to sse2 code so this code may perform better with the
tweeks which include unrolling the inner loop which eliminated the need to use the stack
to hold the counter for one of the loops and replaced incrementing the source register
with address offsets.